refactor(config): import unchanged config structure from latest into old versions
