+++ src/icon/config/v1.py
@@ -1,68 +1,27 @@
-from pathlib import Path
-
 from confz import BaseConfig
 from pydantic import BaseModel
+
+from icon.config.latest import (
+    DatabaseConfig,
+    DataConfiguration,
+    DateConfig,
+    HardwareConfig,
+    HealthCheckConfig,
+    ServerConfig,
+)
 
 __version__ = 1
 
 
-class HealthCheckConfig(BaseModel):
-    interval_seconds: float = 10.0
-
-
-class DataConfiguration(BaseModel):
-    results_dir: str = str(Path.cwd() / "output")
-
-
-class ExperimentLibraryConfigV1(BaseModel):
+class ExperimentLibraryConfig(BaseModel):
     dir: str | None = None
     git_repository: str = "https://..."
     update_interval: int = 30
 
 
-class InfluxDBv1Config(BaseModel):
-    host: str = "localhost"
-    port: int = 8087
-    username: str = "admin"
-    password: str = "admin"  # noqa: S105
-    database: str = "testing"
-    measurement: str = "Experiment Parameters"
-    ssl: bool = False
-    verify_ssl: bool = False
-    headers: dict[str, str] = {}
-
-
-class SQLiteConfig(BaseModel):
-    file: str = str(Path.cwd() / "icon.db")
-
-
-class DatabaseConfig(BaseModel):
-    influxdbv1: InfluxDBv1Config = InfluxDBv1Config()
-    sqlite: SQLiteConfig = SQLiteConfig()
-
-
-class DateConfig(BaseModel):
-    timezone: str = "Europe/Zurich"
-
-
-class PreProcessingConfig(BaseModel):
-    workers: int = 2
-
-
-class ServerConfig(BaseModel):
-    port: int = 8004
-    host: str = "0.0.0.0"
-    pre_processing: PreProcessingConfig = PreProcessingConfig()
-
-
-class HardwareConfig(BaseModel):
-    host: str = "localhost"
-    port: int = 6007
-
-
-class ServiceConfigV1(BaseConfig):  # type: ignore[misc]
+class ServiceConfig(BaseConfig):  # type: ignore[misc]
     version: int = __version__
-    experiment_library: ExperimentLibraryConfigV1 = ExperimentLibraryConfigV1()
+    experiment_library: ExperimentLibraryConfig = ExperimentLibraryConfig()
     databases: DatabaseConfig = DatabaseConfig()
     date: DateConfig = DateConfig()
     server: ServerConfig = ServerConfig()
